Add AuthenticationFilter, add RememberMeFilter, validation implemented

+++ src/main/java/app/yarmak/newsportal/controller/concrete/imp/DeleteNews.java
@@ -6,17 +6,28 @@ import app.yarmak.newsportal.bean.News;
 import app.yarmak.newsportal.controller.concrete.Command;
 import app.yarmak.newsportal.service.NewsService;
 import app.yarmak.newsportal.service.ServiceProvider;
+import app.yarmak.newsportal.util.Validator;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 public class DeleteNews implements Command {
 	private final NewsService newsService = ServiceProvider.getInstance().getNewsService();
+	/**
+	 * Executes the delete news.
+	 */
 
 	@Override
-	public void execute(HttpServletResponse response, HttpServletRequest request) throws ServletException, IOException {
+	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		try {
-			int id = Integer.parseInt(request.getParameter("id"));
+			String idParameter = request.getParameter("id");
+			if (Validator.isNullOrEmpty(idParameter) && Validator.isValidId(idParameter)) {  
+				request.setAttribute("errorMessage", "Идентификатор новости не указан."); 
+				request.getRequestDispatcher("WEB-INF/jsp/error.jsp").forward(request, response); 
+				return; 
+			}
+			
+			int id = Integer.parseInt(idParameter);
 			News news = new News(id, "", "","", "", null, 0, 0, 0,"deleted"); 
 			newsService.deleteNews(news);
 			response.sendRedirect("goController?command=go_to_index_main");
